Learn function added to the network, it is not dynamic yet.

+++ neuron.h
@@ -9,7 +9,7 @@ typedef struct neuron neuron;
 struct neuron
 {
     int nbEntries;
-    int *entries;
+    double *entries;
     double *weight;
     double output;
     double epsilon;
